My portfolio website
This is the start to my online portfolio website. It consists of a home
page which is an about section, a "mywork" page which is where examples
of my work/websites will be located. and A contact page that once a
submission is made on this page you will be redirected to a thank you
page.